% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendTooltips.R
\name{appendTooltips}
\alias{appendTooltips}
\title{construct tooltips for interactive plots}
\usage{
appendTooltips(df)
}
\arguments{
\item{df}{data.frame as returned from `validateScenarios()`}
}
\description{
construct tooltips for interactive plots
}
