% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{average_2020}
\alias{average_2020}
\title{Average 2020 to smoothen Covid shock in historical data}
\usage{
average_2020(hist)
}
\arguments{
\item{hist}{reference data as used in ``validateScenarios()``}
}
\description{
Adds a new model for each model in reference data with smoothed 2020 period
and name "<model>_smoothed".
}
