% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkUnits.R
\name{checkUnits}
\alias{checkUnits}
\title{Check variable consistency}
\usage{
checkUnits(data, cfgRow)
}
\arguments{
\item{data}{scenario or reference data for one variable}

\item{cfgRow}{one row of a config file containing the same variable as the
data object}
}
\description{
Test whether unit of on row of config and data for this variable match.
}
