% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertSCIcriteria.R
\name{convertSCIcriteria}
\alias{convertSCIcriteria}
\title{Import SCI criteria and convert to config}
\usage{
convertSCIcriteria(version = "2026.8.3")
}
\arguments{
\item{version}{criteria version that should be converted to config}
}
\description{
Use installed "scenario-evaluation-criteria" package to convert threshold
criteria from Scenario Compass release to a piamValidation config and a
reference data .rds file
}
