% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importFunctions.R
\name{getConfig}
\alias{getConfig}
\title{import a config shipped with the package}
\usage{
getConfig(config)
}
\arguments{
\item{config}{name of validation config from "inst/config"}
}
\description{
get a validation config file either from "inst/config" (.csv) or by providing
a full path (.csv or .xlsx) or a tibble with the necessary columns
see README or vignette for rules on how to fill the config
}
