% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linePlotThresholds.R
\name{linePlotThresholds}
\alias{linePlotThresholds}
\title{takes the output of "validateScenarios()" and plots heat maps per variable}
\usage{
linePlotThresholds(df, var, reg, scenData = FALSE, refData = FALSE)
}
\arguments{
\item{df}{data.frame to be plotted, as returned by ``validateScenarios()``
plus optional filtering. Only works for datapoint with "reference"}

\item{var}{choose variable to plot}

\item{reg}{choose region to plot}

\item{scenData}{decide whether scenario data should be part of the plot}

\item{refData}{decide whether reference data should be part of the plot}
}
\description{
takes the output of "validateScenarios()" and plots heat maps per variable
}
