% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{listConfigs}
\alias{listConfigs}
\title{List available configs}
\usage{
listConfigs()
}
\value{
character vector of config names
}
\description{
List all validation configuration files that are delivered with the package
and can be directly imported with ``getConfig()`` or used in
``validateScenarios()`` and ``validationReport()``.
}
