% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{listReports}
\alias{listReports}
\title{List available reports}
\usage{
listReports()
}
\value{
character vector of report names
}
\description{
List all validation report templates (.Rmd) that are delivered with the
package and can be used as ``report`` in ``validationReport()``.
}
