% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineData.R
\name{refineRefData}
\alias{refineRefData}
\title{evaluate the content of the "ref_<type>" column and filter
reference data accordingly.
cases:
 - mode chosen
   - range
   - mean
 - no mode chosen
   - use mean if multiple references}
\usage{
refineRefData(ref_data, cfgRow, ref_type = "ref_model")
}
\arguments{
\item{ref_data}{pre-filtered reference data}

\item{cfgRow}{row of validation config used for this data slice}

\item{ref_type}{historical, model, scenario, period}
}
\description{
returns df without variable, unit and <type> columns (see below)
returns df with ref_value_min/max, ref_model, ref_scenario, ref_period
}
