% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{showInputSummary}
\alias{showInputSummary}
\title{Print a summary of the input data}
\usage{
showInputSummary(scen, hist, config)
}
\arguments{
\item{scen}{scenario data as used in ``validateScenarios()``}

\item{hist}{historical/reference data as used in ``validateScenarios()``}

\item{config}{config as used in ``validateScenarios()``}
}
\description{
Print multiple metrics of all data objects given to ``validateScenarios()``
which might be helpful to spot and fix data inconsistencies.
}
