% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateScenarios.R
\name{validateScenarios}
\alias{validateScenarios}
\title{performs the validation checks from a config on a scenario data set}
\usage{
validateScenarios(
  dataPath,
  config,
  outputFile = NULL,
  extraColors = TRUE,
  giveSummary = FALSE
)
}
\arguments{
\item{dataPath}{one or multiple path(s) to scenario data in .mif or .csv
format, in case of historic comparison, also path to reference data}

\item{config}{select config from inst/config or give a full path to a config
file on your computer}

\item{outputFile}{give name of output file in case results should be exported;
include file extension}

\item{extraColors}{if TRUE, use cyan and blue for violation of min thresholds
instead of using the same colors as for max thresholds (yel and red)}

\item{giveSummary}{print a summary of input data via ``showInputSummary()``
which allows spotting data inconsistencies}
}
\description{
performs the validation checks from a config on a scenario data set
}
