% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationHeatmap.R
\name{validationHeatmap}
\alias{validationHeatmap}
\title{takes the output of "validateScenarios()" and plots heatmaps per variable}
\usage{
validationHeatmap(
  df,
  var,
  met,
  historical = TRUE,
  interactive = TRUE,
  compareModels = FALSE
)
}
\arguments{
\item{df}{data.frame as returned by ``validateScenarios()``
and ``appendTooltips()``}

\item{var}{variable to be plotted}

\item{met}{choose metric from "relative", "difference", "absolute" or
"growthrate"}

\item{historical}{should this be a plot comparing to historical data}

\item{interactive}{return plots as interactive plotly plots by default}

\item{compareModels}{if TRUE, plots compare models instead of scenarios}
}
\description{
takes the output of "validateScenarios()" and plots heatmaps per variable
}
