% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputFunctions.R
\name{validationPass}
\alias{validationPass}
\title{returns information on whether scenarios passed critical validation checks}
\usage{
validationPass(data, yellowFail = FALSE)
}
\arguments{
\item{data}{data.frame as returned from ``validateScenarios()``}

\item{yellowFail}{if set to TRUE a yellow check result of a critical
variable will lead to the scenario not passing as validated}
}
\description{
returns information on whether scenarios passed critical validation checks
}
