% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importFunctions.R
\name{importScenarioData}
\alias{importScenarioData}
\title{import IAM data for validation}
\usage{
importScenarioData(scenarioPath)
}
\arguments{
\item{scenarioPath}{one or multiple paths to .mif, .csv, .rds or .xlsx file(s)
or a data.frame containing scenario data in IAM format}
}
\description{
import IAM data for validation
}
