% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linePlotThresholds.R
\name{linePlotThresholds}
\alias{linePlotThresholds}
\title{takes the output of "validateScenarios()" creates a line plot}
\usage{
linePlotThresholds(
  valiData,
  scenData = NULL,
  refData = NULL,
  xlim = c(2010, 2030),
  interactive = TRUE
)
}
\arguments{
\item{valiData}{data to be plotted, as returned by ``validateScenarios()``
and after filtering for one variable and one region.}

\item{scenData}{hand over additional scenario data to be plotted alongside
the validation data. Will use the same variable and region, otherwise
all available data.}

\item{refData}{hand over additional reference data to be plotted alongside
the validation data. Will use the same variable and region, otherwise
all available data.}

\item{xlim}{set limits for the x axis}

\item{interactive}{decide if an interactive ggploty object should be returned}
}
\description{
takes the output of "validateScenarios()" creates a line plot
}
