% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piamValidation-package.R
\docType{package}
\name{piamValidation-package}
\alias{piamValidation-package}
\alias{piamValidation}
\title{Validation Tools for PIK-PIAM}
\description{
The piamValidation package provides validation tools for the Potsdam Integrated Assessment Modelling environment.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/piamValidation}
  \item \url{https://pik-piam.github.io/piamValidation}
}

}
\author{
\strong{Maintainer}: Pascal Weigmann \email{pascal.weigmann@pik-potsdam.de}

Authors:
\itemize{
  \item Oliver Richters
  \item Fabrice Lécuyer
}

}
