% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piamValidation-package.R
\docType{package}
\name{piamValidation-package}
\alias{piamValidation-package}
\alias{piamValidation}
\title{Validation Tools for PIK-PIAM}
\description{
The piamValidation package provides validation tools for the Potsdam Integrated Assessment Modelling environment.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/piamValidation}
  \item \url{https://pik-piam.github.io/piamValidation}
}

}
\author{
\strong{Maintainer}: Pascal Weigmann \email{pascal.weigmann@pik-potsdam.de} (\href{https://orcid.org/0000-0001-8801-173X}{ORCID}) (Potsdam Institute for Climate Impact Research)

Authors:
\itemize{
  \item Oliver Richters (\href{https://orcid.org/0000-0001-8253-4716}{ORCID}) (Potsdam Institute for Climate Impact Research)
  \item Fabrice Lécuyer (\href{https://orcid.org/0000-0002-7364-999X}{ORCID}) (Potsdam Institute for Climate Impact Research)
  \item Johannes Koch (\href{https://orcid.org/0000-0003-2920-8086}{ORCID}) (Potsdam Institute for Climate Impact Research)
}

}
