% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateScenarios.R
\name{validateScenarios}
\alias{validateScenarios}
\title{performs the validation checks from a config on a scenario dataset}
\usage{
validateScenarios(scenarioPath, configName)
}
\arguments{
\item{scenarioPath}{one or multiple path(s) to scenario data in .mif or .csv
format, in case of historic comparison, append path to ref data}

\item{configName}{select config from inst/config}
}
\description{
performs the validation checks from a config on a scenario dataset
}
