% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateScenarios.R
\name{validateScenarios}
\alias{validateScenarios}
\title{performs the validation checks from a config on a scenario dataset}
\usage{
validateScenarios(scenarioPath, configName, referencePath = NULL)
}
\arguments{
\item{scenarioPath}{one or multiple path(s) to scenario data in .mif or .csv
format}

\item{configName}{select config from inst/config}

\item{referencePath}{in case of historic comparison, choose path to ref data}
}
\description{
performs the validation checks from a config on a scenario dataset
}
