% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationHeatmap.R
\name{validationHeatmap}
\alias{validationHeatmap}
\title{takes the output of "validateScenarios()" and plots heatmaps per variable}
\usage{
validationHeatmap(data, var, cat, met, interactive = T, compareModels = T)
}
\arguments{
\item{data}{data.frame as returned by ``validateScenarios()``}

\item{var}{variable to be plotted}

\item{cat}{choose category from "historical" or "scenario"}

\item{met}{choose metric from "relative", "difference", "absolute" or
"growthrate"}

\item{interactive}{return plots as interactive plotly plots by default}

\item{compareModels}{if TRUE, plots compare models instead of scenarios}
}
\description{
takes the output of "validateScenarios()" and plots heatmaps per variable
}
