% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryHeatmap.R, R/validationHeatmap.R
\name{validationHeatmap}
\alias{validationHeatmap}
\title{takes the output of "validateScenarios()" and plots heatmaps per variable}
\usage{
validationHeatmap(
  df,
  var,
  met,
  historical = TRUE,
  interactive = TRUE,
  x_plot = "region",
  y_plot = "period",
  x_facet = "model",
  y_facet = "scenario"
)

validationHeatmap(
  df,
  var,
  met,
  historical = TRUE,
  interactive = TRUE,
  x_plot = "region",
  y_plot = "period",
  x_facet = "model",
  y_facet = "scenario"
)
}
\arguments{
\item{df}{data.frame as returned by ``validateScenarios()``
and ``appendTooltips()``}

\item{var}{variable to be plotted}

\item{met}{choose metric from "relative", "difference", "absolute" or
"growthrate"}

\item{historical}{should this be a plot comparing to historical data}

\item{interactive}{return plots as interactive plotly plots by default}

\item{x_plot}{choose dimension to display on x-axis of plot, default: region}

\item{y_plot}{choose dimension to display on y-axis of plot, default: period}

\item{x_facet}{choose dimension to display on x-dim of facets, default: model}

\item{y_facet}{choose dimension to display on x-dim of facets, default: scenario}
}
\description{
takes the output of "validateScenarios()" and plots heatmaps per variable

takes the output of "validateScenarios()" and plots heatmaps per variable
}
