% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationHeatmap.R
\name{validationHeatmap}
\alias{validationHeatmap}
\title{takes the output of "validateScenarios()" and plots heat maps per variable}
\usage{
validationHeatmap(
  valiData,
  main_dim = "variable",
  x_plot = NULL,
  y_plot = NULL,
  x_facet = NULL,
  y_facet = NULL,
  interactive = TRUE
)
}
\arguments{
\item{valiData}{data to be plotted, as returned by ``validateScenarios()``
(and ``appendTooltips()`` if interactive), plus optional filtering.
Needs to have at least one dimension with only one unique element.}

\item{main_dim}{out of the 5-dim df, 1 dim has to contain only on element,
this is the main dimension of the plot, default: variable}

\item{x_plot}{choose dimension to display on x-axis of plot, if any
is NULL, arrangement is chosen automatically based on data dimensions}

\item{y_plot}{choose dimension to display on y-axis of plot}

\item{x_facet}{choose dimension to display on x-dim of facets}

\item{y_facet}{choose dimension to display on x-dim of facets}

\item{interactive}{return plots as interactive plotly plots by default}
}
\description{
takes the output of "validateScenarios()" and plots heat maps per variable
}
