% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationReport.R
\name{validationReport}
\alias{validationReport}
\title{perform validateScenarios and create an .html report using .Rmd templates}
\usage{
validationReport(dataPath, config, report = "default")
}
\arguments{
\item{dataPath}{one or multiple path(s) to scenario data in .mif or .csv
format}

\item{config}{name of a config from inst/config}

\item{report}{specify which .Rmd file should be used to create report}
}
\description{
perform validateScenarios and create an .html report using .Rmd templates
}
